update backend files

+++ backend/src/main/java/pika/hungt1/dx/repository/SalesOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pika.hungt1.dx.entity.SalesOrder;
 
 @NullMarked
-public interface SalesOrderRepository extends JpaRepository<SalesOrder, Long> {
+public interface SalesOrderRepository extends JpaRepository<SalesOrder, Integer> {
 }
